requirements.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pbr==3.1.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.2
rope==0.14.0
six==1.10.0
sqlparse==0.3.0
stevedore==1.25.0
typed-ast==1.4.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wrapt==1.11.2
